Makefile: update release-hack
We must not prevent from setting "VERSION_REVISION" to have our intented
text in the REVISION-field. This was needed till v0.3.0 only.
Alfter 0.3.0 release this was changed to "VERSION_CODE" and this value
can be found in "/etc/openwrt_revision".
To not include git-revision in filename of releases we simply set
CONFIG_VERSION_CODE_FILENAMES=n when BUILD_TYPE=release